Update to latest i.MX 8M Plus SDK 2.16.0

+++ middleware/multicore/rpmsg_lite/lib/include/platform/imx8mp_m7/rpmsg_platform.h
@@ -1,7 +1,5 @@
 /*
- * Copyright 2019,2022 NXP
- * All rights reserved.
- *
+ * Copyright 2019-2023 NXP
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -50,8 +48,8 @@ void platform_time_delay(uint32_t num_msec);
 void platform_map_mem_region(uint32_t vrt_addr, uint32_t phy_addr, uint32_t size, uint32_t flags);
 void platform_cache_all_flush_invalidate(void);
 void platform_cache_disable(void);
-uint32_t platform_vatopa(void *addr);
-void *platform_patova(uint32_t addr);
+uintptr_t platform_vatopa(void *addr);
+void *platform_patova(uintptr_t addr);
 
 /* platform init/deinit */
 int32_t platform_init(void);
